+++ src/main/java/baseball/Application.java
@@ -3,6 +3,8 @@ package baseball;
 public class Application {
     public static void main(String[] args) {
         //TODO: 숫자 야구 게임 구현
+
+        //새로운 게임 생성 후 시작
         new PlayBall().startGame();
     }
 }