+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 1. ```
    2021/11/24
-   "feat : add README.md to list up application functions"
+   "docs(readme.md) : add README.md to list up application functions"
    
    프로젝트 생성 및 요구사항에 맞는 기능 목록을 포함하여 리드미 작성
    ```
@@ -22,14 +22,23 @@
    ```
 
 3. ```
-   2021/11/24
-   "fix : change number pacakge structure to seperate model and service"
+   2021/11/25
+   "refactor : change number pacakge structure to seperate model and service"
    
    컴퓨터 난수 생성과 컴퓨터 난수 저장의 역할 분리를 위한 코드 리팩토링
    baseball이라는 인터페이스로 통합하여 추후 추가될 사용자 입력에 대한 처리도 가능하게 함
    ```
 
-4. 
+4. ```
+   2021/11/25
+   "feat : add user input view and service with handling exception"
+   
+   사용자 입력에 관한 기능 추가
+   입력하기 위한 출력 메세지 추가
+   입력을 판단하고 이에 따른 어플리케이션 진행여부를 결정 하도록 함
+   ```
+
+5.